*: refactor sendto action and HTTPMutator
